1，无缝大地图功能的开发未调试。2，整合mongodb的第三方库。

+++ core/src/lib/common/timesys.c
@@ -53,7 +53,7 @@ unsigned int GetCurrentSec()
 #ifdef _WIN32
 	struct timeb rawtime;
 	ftime(&rawtime);
-	return rawtime.time;
+	return (unsigned int)rawtime.time;
 #else
 	struct timeval tv;
 	gettimeofday(&tv, NULL); //Under Linux, this function is in the sys / time. H header file
